requirements.txt
colorama~=0.4.5
pytube~=12.1.0
moviepy~=1.0.3
progressbar2~=4.0.0